REFACTOR(DO): change table `user` ddl

+++ src/main/java/chen/binflare/dao/UserDAO.java
@@ -3,5 +3,5 @@ package chen.binflare.dao;
 import chen.binflare.entity.UserDO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDAO extends JpaRepository<UserDO, Integer> {
+public interface UserDAO extends JpaRepository<UserDO, Long> {
 }
